+++ tdconvert/tests/test_library.py
@@ -23,16 +23,14 @@ if __name__ == "__main__":
     print_separator()
 
     # Test
-    print_subheader(f"Testing Random Flag")
+
+    print_subheader(f"Testing Sample Data")
 
     profile = "test"
     overrides = None
     TestHub = tdconvert.properties.initialize_app(profile, overrides)
-    database_populated = tdconvert.library.import_sample_data(TestHub)
-    if database_populated:
-        flag_row = tdconvert.library.get_random_flag(TestHub)
-        print(flag_row["flag_name"])
-        print()
+    database_populated = tdconvert.library.populate_sample_data(TestHub)
+    print(f"Database populated: {database_populated}")
 
     print_separator()